Add SATTP with guesses; closes #796

+++ wrappers/Python/CoolProp/AbstractState.pxd
@@ -12,9 +12,9 @@ cdef class PyPhaseEnvelopeData:
     cpdef public list T, p, lnT, lnp, rhomolar_liq, rhomolar_vap, lnrhomolar_liq, lnrhomolar_vap, hmolar_liq, hmolar_vap, smolar_liq, smolar_vap, Q
     
 cdef class PyGuessesStructure:
-    cpdef public long double T, p, rhomolar, hmolar, smolar
-    cpdef public long double rhomolar_liq, rhomolar_vap
-    cpdef public vector[long double] x, y
+    cpdef public double T, p, rhomolar, hmolar, smolar
+    cpdef public double rhomolar_liq, rhomolar_vap
+    cpdef public vector[double] x, y
 
 cdef class AbstractState:
     cdef cAbstractState.AbstractState *thisptr     # hold a C++ instance which we're wrapping
